Move landing page from routes folder to components folder
