+++ met-web/src/routes/AuthenticatedRoutes.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import NotFound from './NotFound';
-import LandingPage from './LandingPage/LandingPage';
+import LandingPage from '../components/LandingPage/LandingPage';
 import { ThemeProvider } from '@mui/system';
 import { BaseTheme, PublicTheme } from '../styles/Theme';
 import UserService from '../services/userService';
